requirements.txt
alembic==1.15.2
asttokens==3.0.0
catboost==1.2.8
category_encoders==2.8.1
cloudpickle==3.1.1
colorlog==6.9.0
colour==0.1.5
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dtreeviz==2.2.2
executing==2.2.0
fonttools==4.58.0
graphviz==0.20.3
greenlet==3.2.2
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.0
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mljar-scikit-plot==0.3.12
mljar-supervised==1.1.17
narwhals==1.39.1
numba==0.61.2
numpy==1.26.4
nvidia-nccl-cu12==2.26.5
optuna==4.3.0
optuna-integration==4.3.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
plotly==6.1.0
pluggy==1.6.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shap==0.47.2
six==1.17.0
slicer==0.0.8
SQLAlchemy==2.0.41
stack-data==0.6.3
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
wordcloud==1.9.4
xgboost==3.0.1
